group the api endpoints that contact OP in an OCS controller
